Added Sweet Alerts

+++ pages/validated.js
@@ -2,6 +2,8 @@ import { useAccount } from "@components/hooks/web3/useAccount"
 import { AdminLayout } from "@components/layout"
 import { EditPropertyModal, PropertyCard, PropertyList } from "@components/properties"
 import { useState, useEffect } from "react"
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
 
 /**
  * Displays valiadted properties
@@ -22,6 +24,9 @@ export default function Validated() {
     // Sets validator constant to MetaMask account address
     const validator = account.data
 
+    const MakeAvailableSwal = withReactContent(Swal)
+    const DeleteSwal = withReactContent(Swal)
+
     /**
      * Fetch properties from database where property 
      * validator == current MetaMask account address
@@ -39,28 +44,59 @@ export default function Validated() {
      * Change the status of the selected property 
      * from "Created" to "Pending"
      * 
-     * @param   {object}    _property Selected property
+     * @param   {object}    property Selected property
      */
-    const handleMakeAvailable = (_property) => {
-        fetch("http://localhost:3500/properties/" + _property.id, {
+    const handleMakeAvailable = async (e, property) => {
+        e.preventDefault()
+        fetch("http://localhost:3500/properties/" + property.id, {
             method: "PATCH",
             headers: { "Content-Type": "application/json" },
             body: JSON.stringify({
-                ..._property,
+                ...property,
                 status: "Pending"
             })
-        }).then(location.reload())
+        })
+        await MakeAvailableSwal.fire({
+            title: <h3 className="pb-3 text-lg font-bold leading-6 text-gray-900 border-b">Property marked as available!</h3>,
+            confirmButtonColor: "#d33",
+            icon: "success"
+        }).then((result) => {
+            if(result.isConfirmed) {
+                location.reload()
+            }
+        })
     }
 
     /**
      * Deletes the selected property from the database
      * 
-     * @param {object} _property 
+     * @param {object} property 
      */
-    const handleDetele = (_property) => {
-        fetch("http://localhost:3500/properties/" + _property.id, {
-            method: "DELETE"
-        }).then(location.reload())
+    const handleDetele = async (e, property) => {
+        e.preventDefault()
+        DeleteSwal.fire({
+            title: <h3 className="pb-3 text-lg font-bold leading-6 text-gray-900 border-b">Are you sure you want to delete property?</h3>,
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#d33",
+            cancelButtonColor: '#6b7280',
+            confirmButtonText: "Confirm"
+        }).then((result) => {
+            if(result.isConfirmed) {
+                fetch("http://localhost:3500/properties/" + property.id, {
+                    method: "DELETE"
+                })
+                DeleteSwal.fire({
+                    title: <h3 className="pb-3 text-lg font-bold leading-6 text-gray-900 border-b">Property deleted!</h3>,
+                    confirmButtonColor: "#d33",
+                    icon: "success"
+                }).then((result) => {
+                    if(result.isConfirmed) {
+                        location.reload()
+                    }
+                })
+            }
+        })
     }
 
     return (
@@ -84,7 +120,7 @@ export default function Validated() {
                         <button
                             disabled={property.status=="Pending"}
                             className="disabled:opacity-50 disabled:cursor-not-allowed bg-blue-500 hover:bg-blue-600 focus:ring-blue-200 focus:ring-4 text-white font-medium rounded-lg text-sm p-2 text-center"
-                            onClick = {() => handleMakeAvailable(property)}>
+                            onClick = {(e) => handleMakeAvailable(e, property)}>
                             Make Available
                         </button>
                         <button
@@ -98,7 +134,7 @@ export default function Validated() {
                         <button
                             className="disabled:opacity-50 disabled:cursor-not-allowed"
                             disabled={property.status=="Pending"}
-                            onClick = {() => handleDetele(property)}>
+                            onClick = {(e) => handleDetele(e, property)}>
                             <svg className="w-9 h-9 bg-red-500 hover:bg-red-600 focus:ring-red-200 focus:ring-4 text-white rounded-lg p-1" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
                                 <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"></path>
                             </svg>
